+++ passes/S015/S015.go
@@ -10,6 +10,7 @@ import (
 
 	"golang.org/x/tools/go/analysis"
 
+	"github.com/bflad/tfproviderlint/helper/terraformtype"
 	"github.com/bflad/tfproviderlint/passes/commentignore"
 	"github.com/bflad/tfproviderlint/passes/schemamap"
 )
@@ -43,7 +44,7 @@ func run(pass *analysis.Pass) (interface{}, error) {
 			continue
 		}
 
-		for _, attributeName := range schemamap.GetSchemaAttributeNames(smap) {
+		for _, attributeName := range terraformtype.GetSchemaMapAttributeNames(smap) {
 			switch t := attributeName.(type) {
 			default:
 				continue