+++ passes/schemamap/schemamap.go
@@ -31,7 +31,7 @@ func run(pass *analysis.Pass) (interface{}, error) {
 	inspect.Preorder(nodeFilter, func(n ast.Node) {
 		x := n.(*ast.CompositeLit)
 
-		if !isSchemaMap(pass, x) {
+		if !terraformtype.IsMapStringHelperSchemaTypeSchema(x, pass.TypesInfo) {
 			return
 		}
 
@@ -39,62 +39,4 @@ func run(pass *analysis.Pass) (interface{}, error) {
 	})
 
 	return result, nil
-}
-
-func isSchemaMap(pass *analysis.Pass, cl *ast.CompositeLit) bool {
-	switch v := cl.Type.(type) {
-	default:
-		return false
-	case *ast.MapType:
-		switch k := v.Key.(type) {
-		default:
-			return false
-		case *ast.Ident:
-			if k.Name != "string" {
-				return false
-			}
-		}
-
-		switch mv := v.Value.(type) {
-		default:
-			return false
-		case *ast.StarExpr:
-			return terraformtype.IsHelperSchemaTypeSchema(pass.TypesInfo.TypeOf(mv.X))
-		}
-	}
-}
-
-// GetSchemaAttributes returns all attributes held in a map[string]*schema.Schema
-func GetSchemaAttributes(schemamap *ast.CompositeLit) []*ast.CompositeLit {
-	var result []*ast.CompositeLit
-
-	for _, elt := range schemamap.Elts {
-		switch v := elt.(type) {
-		default:
-			continue
-		case *ast.KeyValueExpr:
-			switch v := v.Value.(type) {
-			default:
-				continue
-			case *ast.CompositeLit:
-				result = append(result, v)
-			}
-		}
-	}
-	return result
-}
-
-// GetSchemaAttributeNames returns all attribute names held in a map[string]*schema.Schema
-func GetSchemaAttributeNames(schemamap *ast.CompositeLit) []ast.Expr {
-	var result []ast.Expr
-
-	for _, elt := range schemamap.Elts {
-		switch v := elt.(type) {
-		default:
-			continue
-		case *ast.KeyValueExpr:
-			result = append(result, v.Key)
-		}
-	}
-	return result
 }